remove TemporaryOpenDocumentFile; implement OpenDocumentType; remove OpenDocument.test submodule (will be back soon); renew document handling

+++ src/at/stefl/opendocument/java/util/DefaultFileCache.java
@@ -4,6 +4,7 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
+import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.io.RandomAccessFile;
@@ -37,14 +38,21 @@ public class DefaultFileCache extends AbstractFileCache {
 	return directory;
     }
 
-    @Override
-    public boolean exists(String name) {
-	return getFile(name).exists();
+    private File file(String name) {
+	return new File(directory, name);
     }
 
     @Override
-    public File getFile(String name) {
-	return new File(directory, name);
+    public boolean exists(String name) {
+	return file(name).exists();
+    }
+
+    @Override
+    public File getFile(String name) throws FileNotFoundException {
+	File file = file(name);
+	if (!file.exists())
+	    throw new FileNotFoundException();
+	return file;
     }
 
     @Override
@@ -71,12 +79,24 @@ public class DefaultFileCache extends AbstractFileCache {
     }
 
     @Override
-    public File create(String name) {
-	return getFile(name);
+    public File create(String name) throws IOException {
+	File file = file(name);
+	new FileOutputStream(file).close();
+	return file;
     }
 
     @Override
-    public void delete(String name) {
+    public File move(String from, String to) throws FileNotFoundException {
+	File file = file(to);
+	getFile(from).renameTo(file(to));
+	return file;
+    }
+
+    @Override
+    public void delete(String name) throws FileNotFoundException {
+	File file = getFile(name);
+	if (file == null)
+	    throw new FileNotFoundException();
 	getFile(name).delete();
     }
